requirements.txt
# Python dependencies for BizCopilot Connectors on Vercel
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
mysql-connector-python>=8.2.0
pymongo>=4.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
